Fix 3 critical security vulnerabilities
- Fix uncontrolled format string in command.c:203,231
- Fix wrong type format specifier in register.c:282 (%d -> %ld)
- Fix integer overflow risk in command.c:49 memory allocation

Addresses CodeQL security findings for safer code execution.

+++ register.c
@@ -278,8 +278,8 @@ void editorViewRegister(struct editorConfig *ed,
 				       ed->registers[reg].rdata.region);
 		break;
 	case REGISTER_NUMBER:
-		editorSetStatusMessage("%s (number): %d", str,
-				       ed->registers[reg].rdata.number);
+		editorSetStatusMessage("%s (number): %ld", str,
+				       (long)ed->registers[reg].rdata.number);
 		break;
 	case REGISTER_POINT:;
 		struct editorPoint *pt = ed->registers[reg].rdata.point;
